@@ -30,7 +30,7 @@ ext_modules = [Extension("_foo", ["stub.cc"])] if platform.startswith("linux") e
 setup(
     name="larq-compute-engine",
     version=get_version_number(default="0.11.1"),
-    python_requires=">=3.7",
+    python_requires=">=3.9",
     description="Highly optimized inference engine for binarized neural networks.",
     long_description=readme(),
     long_description_content_type="text/markdown",
@@ -39,12 +39,7 @@ setup(
     packages=find_packages(),
     ext_modules=ext_modules,
     url="https://larq.dev/",
-    install_requires=[
-        "flatbuffers>=2.0",
-        "packaging>=19",
-        "tqdm>=4",
-        "importlib-metadata ~= 2.0 ; python_version<'3.8'",
-    ],
+    install_requires=["flatbuffers>=2.0", "tqdm>=4"],
     extras_require={
         "tensorflow": ["tensorflow>=1.14"],
         "tensorflow_gpu": ["tensorflow-gpu>=1.14"],
@@ -60,10 +55,9 @@ setup(
         "License :: OSI Approved :: Apache Software License",
         "Programming Language :: Python :: 3",
         "Programming Language :: Python :: 3 :: Only",
-        "Programming Language :: Python :: 3.7",
-        "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
         "Topic :: Scientific/Engineering",
         "Topic :: Scientific/Engineering :: Mathematics",
         "Topic :: Scientific/Engineering :: Artificial Intelligence",
